requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic[email]>=2.7.4
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9

# DB
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1

# Graph
neo4j>=5.20.0

# Vector
weaviate-client>=4.6.1

# Cache
redis>=5.0.4

# LLM + orchestration
langchain>=0.2.3
langchain-openai>=0.1.17
langchain-anthropic>=0.1.15
langsmith>=0.1.112

# Frontend
streamlit>=1.35.0

# Utils
python-dotenv>=1.0.1
httpx>=0.27.0
tenacity>=8.3.0

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0